+++ frontend/src/components/MovieItem.jsx
@@ -6,6 +6,7 @@ const { Meta } = Card;
 export default function MovieItem({movie}) {
     const imageUrl  =  movie.posters?.split('|')[0];
     const {title, nation, openDt} = movie || {};
+    const {movieId, plot} = movie || {};
 
     return (
         <Card
