+++ frontend/src/layout/Header.jsx
@@ -1,27 +1,25 @@
 import React from "react";
-import {Link} from "react-router-dom";
-import { Input, Space } from 'antd';
+import axios from "axios";
+import SearchBar from "../components/SearchBar";
 
-const { Search } = Input;
+export default function Header({setMovies}) {
+    const baseUrl = "http://localhost:8080";
 
-function Header({setSearchKeyword, handleSearchClick}) {
+    async function handleSearchClick(searchKeyword) {
+        try {
+            const searchResponse = await axios.get( baseUrl + "/api/search", {
+                params : {
+                    keyWord : searchKeyword
+                }
+            });
+            setMovies(searchResponse.data);
+
+        } catch (e) {
+            console.error(e);
+        }
+    }
 
     return (
-        <header>
-            <div className="App-logo">
-                <a href="/">LAHO-PEDIA</a>
-            </div>
-            <nav className="Nav">
-                <ul>
-                    <li><Link to="/myPage">마이페이지</Link></li>
-                    <li><Link to="/login">로그인</Link></li>
-                    <li>
-                        <Search placeholder="검색어를 입력하세요." onSearch={handleSearchClick} style={{ width: 200 }} />
-                    </li>
-                </ul>
-            </nav>
-        </header>
-    );
+        <SearchBar handleSearchClick = {handleSearchClick}/>
+    )
 }
-
-export default Header;
